requirements.txt
numpy
pandas
matplotlib
scikit-learn
requests
python-dotenv
seaborn
scipy
joblib
python-dotenv
setuptools